+++ treasure-api/src/main/java/io/treasure/service/UserCouponService.java
@@ -6,12 +6,13 @@ import io.treasure.entity.MerchantCouponEntity;
 import io.treasure.entity.UserCouponEntity;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserCouponService extends CrudService<UserCouponEntity, UserCouponDTO> {
 
     List selectByUserId(Long userId, long martId, double money);
-    double  selectGift(Long userId);
+    BigDecimal selectGift(Long userId);
     UserCouponEntity   selectByCouponId(long couponId,long userId);
     List  selectMyCouponByUserId(long userId);
     List<MerchantCouponEntity>  selectMartCoupon(Long userId, long martId);
